fixed build app
